pyproject.toml
[project]
name = "journalist-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "feedparser>=6.0.12",
    "openrouter>=0.9.1",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.12",
    "ty>=0.0.32",
]
